requirements.txt
certifi==2019.11.28
cffi==1.14.0
cryptography==2.8
databases==0.2.6
decorator==4.4.2
future==0.18.2
orm==0.1.5
peewee==3.13.1
pycparser==2.20
python-decouple==3.3
python-telegram-bot==12.4.2
six==1.14.0
SQLAlchemy==1.3.13
tornado==6.0.4
typesystem==0.2.4